split countdown page folder into two
