+++ client/slices/block/pages/countdown-details/BlockCountdownTimer.tsx
