+++ client/slices/block/pages/countdown-details/__screenshots__/BlockCountdown.pw.tsx_default_long-period-until-the-block-mobile-base-view-1.png
