+++ client/slices/block/pages/countdown-details/__screenshots__/BlockCountdown.pw.tsx_default_short-period-until-the-block-desktop-1.png
