+++ client/slices/block/pages/countdown-details/create-google-calendar-link.ts
