+++ client/slices/block/pages/countdown-details/create-ics-file-blob.ts
