+++ client/slices/block/pages/countdown-details/split-seconds-in-periods.ts
