+++ client/slices/block/pages/countdown-index/__screenshots__/BlockCountdownIndex.pw.tsx_default_base-view-mobile-1.png
