+++ client/slices/block/pages/countdown-index/__screenshots__/BlockCountdownIndex.pw.tsx_mobile_base-view-mobile-1.png
